+++ internal/pkg/deffile/deffile.go
@@ -16,11 +16,11 @@ import (
 
 	"github.com/gvallee/go_util/pkg/util"
 	"github.com/sylabs/singularity-mpi/internal/pkg/distro"
-	"github.com/sylabs/singularity-mpi/internal/pkg/implem"
 	"github.com/sylabs/singularity-mpi/internal/pkg/ldd"
 	"github.com/sylabs/singularity-mpi/pkg/app"
 	"github.com/sylabs/singularity-mpi/pkg/buildenv"
 	"github.com/sylabs/singularity-mpi/pkg/container"
+	"github.com/sylabs/singularity-mpi/pkg/implem"
 	"github.com/sylabs/singularity-mpi/pkg/sys"
 )
