+++ internal/pkg/deffile/deffile_test.go
@@ -13,9 +13,9 @@ import (
 	"testing"
 
 	"github.com/sylabs/singularity-mpi/internal/pkg/distro"
-	"github.com/sylabs/singularity-mpi/internal/pkg/implem"
 	"github.com/sylabs/singularity-mpi/pkg/app"
 	"github.com/sylabs/singularity-mpi/pkg/buildenv"
+	"github.com/sylabs/singularity-mpi/pkg/implem"
 	"github.com/sylabs/singularity-mpi/pkg/sys"
 )
