+++ pkg/jm/jobmgr_native.go
@@ -13,10 +13,10 @@ import (
 	"strconv"
 
 	"github.com/sylabs/singularity-mpi/internal/pkg/impi"
-	"github.com/sylabs/singularity-mpi/internal/pkg/implem"
 	"github.com/sylabs/singularity-mpi/internal/pkg/job"
-	"github.com/sylabs/singularity-mpi/internal/pkg/mpi"
 	"github.com/sylabs/singularity-mpi/pkg/buildenv"
+	"github.com/sylabs/singularity-mpi/pkg/implem"
+	"github.com/sylabs/singularity-mpi/pkg/mpi"
 	"github.com/sylabs/singularity-mpi/pkg/syexec"
 	"github.com/sylabs/singularity-mpi/pkg/sys"
 )
