+++ pkg/jm/jobmgr_slurm_test.go
